Fix int8 conversions
I just started reading, and the very first example was int8(128) getting coerced to 127, which kinda blew my mind. Then it described it as a C-style cast, which really confused things. I tried it in Firefox, and I got the expected -128. The closest thing I could think of that would do anything remotely like this is new Uint8ClampedArray([256]). I guess this is just a typo?